# IL-based serializer improvements