--- a/src/Orleans/Orleans.csproj
+++ b/src/Orleans/Orleans.csproj
@@ -136,9 +136,7 @@
     <Compile Include="Runtime\IGrainRuntime.cs" />
     <Compile Include="Serialization\IlBasedFallbackSerializer.cs" />
     <Compile Include="Serialization\IlCodeGenerationException.cs" />
-    <Compile Include="Serialization\IlBasedSerializers.cs" />
     <Compile Include="Serialization\IlBasedSerializerGenerator.cs" />
-    <Compile Include="Serialization\IlBasedSerializerTypeChecker.cs" />
     <Compile Include="Serialization\IlDelegateBuilder.cs" />
     <Compile Include="Serialization\OrleansJsonSerializer.cs" />
     <Compile Include="Serialization\BinaryFormatterSerializer.cs" />