--- a/src/Orleans/Orleans.csproj
+++ b/src/Orleans/Orleans.csproj
@@ -134,10 +134,10 @@
     <Compile Include="IDs\UniqueKey.cs" />
     <Compile Include="Core\IGrainFactory.cs" />
     <Compile Include="Runtime\IGrainRuntime.cs" />
-    <Compile Include="Serialization\IlBasedFallbackSerializer.cs" />
-    <Compile Include="Serialization\IlCodeGenerationException.cs" />
-    <Compile Include="Serialization\IlBasedSerializerGenerator.cs" />
-    <Compile Include="Serialization\IlDelegateBuilder.cs" />
+    <Compile Include="Serialization\ILBasedSerializer.cs" />
+    <Compile Include="Serialization\ILGenerationException.cs" />
+    <Compile Include="Serialization\ILSerializerGenerator.cs" />
+    <Compile Include="Serialization\ILDelegateBuilder.cs" />
     <Compile Include="Serialization\OrleansJsonSerializer.cs" />
     <Compile Include="Serialization\BinaryFormatterSerializer.cs" />
     <Compile Include="Serialization\ReflectedSerializationMethodInfo.cs" />