… src/Orleans/Serialization/IlBasedSerializerGenerator.cs
… src/Orleans/Serialization/IlBasedSerializerTypeChecker.cs
… src/Orleans/Serialization/IlBasedSerializers.cs
… src/Orleans/Serialization/SerializationContext.cs
… src/Orleans/Utils/ReferenceEqualsComparer.cs
… test/Benchmarks/OrleansBenchmarks/Serialization/SerializationBenchmarks.cs
… test/NonSiloTests/Orleans.NonSiloTests.csproj
… test/NonSiloTests/Serialization/BuiltInSerializerTests.cs
… test/NonSiloTests/Serialization/ILBasedSerializerTests.cs
… vNext/src/Orleans/Orleans.csproj
… vNext/src/Orleans/project.json
… vNext/test/Tester/project.json