Hopefully the way these commits are split out makes it fairly easy to follow what's going on.

* Efficient enum serialization (statically selects the right integral type & directly encodes `stream.Write` & `Enum.ToObject` calls)
* Direct (non-boxing) serialization of simple types 
* Improvements to emitted IL:
  * Select between `call` & `callvirt` OpCode depending on method
  * Select between `Ldsfld` & `Ldfld` depending on whether a field is static or non-static
  * Use optimal OpCodes where applicable. Eg: `Ldarg_0` with no `int` value rather than `Ldarg` + `0`
* Cache result of `Type.GetTypeFromHandle` in static fields for generated IL rather than calling that method every time a `Type` is needed.
* (more) Compact representation of types
* Misc fixes like consolidating & renaming classes + reducing allocations

Here're the perf numbers before & after this change using the `FallbackFullLoop` benchmark method.

**Before:**

        Serializer |      Median |    StdDev | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------------------ |------------ |---------- |------ |------ |------ |------------------- |
           Default | 284.7841 us | 9.0201 us | 36.00 |     - |     - |          34,755.93 |
 ILBasedSerializer |  50.5088 us | 1.5490 us |  1.38 |     - |     - |           3,295.73 |

**After: 3x quicker and allocates 1/3 as much memory**

        Serializer |      Median |    StdDev | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------------------ |------------ |---------- |------ |------ |------ |------------------- |
 ILBasedSerializer |  17.3400 us | 0.3762 us | 18.00 |     - |     - |           1,310.49 |


## Further work which could be done (but maybe isn't worthwhile yet):

* In-place deserialization: instead of DeserializeInner returning an `object`, make it return `void` & take a `ref` to the destination. Avoids copying & boxing.
* Avoid boxing by making `SerializeInner` generic.
* Improve type serialization by giving commonly used types well-known short identifiers instead of serializing type names.