@ReubenBond 2016-10-21
Functional tests pass