@ReubenBond 2016-10-25
@jdom yep, the perf numbers are in the original post. It's 3x faster and allocates 1/3 as much memory. Pushed fixes - Tester's project.json should be pulling the right deps now.